requirements.txt

certifi==2018.4.16
chardet==3.0.4
coreapi==2.3.3
django==2.0.7
django-filter==1.1.0
djangorestframework==3.8.2
markdown==2.6.11
mysqlclient
django-cors-headers
openpyxl
django-redis
redis
arrow==0.15.0
demjson
django_crontab
geoip2
pyotp
cryptokit
cos-python-sdk-v5
tencentcloud-sdk-python
qcloud-python-sts
xmltodict